Make return value of DataInputStream::read_until/read_upto a Vec<u8> and don't return redundant string length
The functions don't necessarily return valid UTF8 strings and as such
must return a plain byte array. This also takes care of nullability as
these functions can return NULL, which would automatically be converted
to an empty Vec now instead of panicking.

In Rust we already have the length as part of the Vec type itself, no
need to keep it around separately.

Fixes https://github.com/gtk-rs/gio/issues/276
