+++ src/auto/data_input_stream.rs
@@ -13,12 +13,10 @@ use glib::GString;
 use glib::StaticType;
 use glib::ToValue;
 use glib_sys;
-use gobject_sys;
 use std::boxed::Box as Box_;
 use std::fmt;
 use std::mem;
 use std::mem::transmute;
-use std::pin::Pin;
 use std::ptr;
 use BufferedInputStream;
 use Cancellable;
@@ -139,56 +137,6 @@ pub trait DataInputStreamExt: 'static {
 
     fn read_uint64<P: IsA<Cancellable>>(&self, cancellable: Option<&P>)
         -> Result<u64, glib::Error>;
-
-    #[cfg_attr(feature = "v2_56", deprecated)]
-    fn read_until<P: IsA<Cancellable>>(
-        &self,
-        stop_chars: &str,
-        cancellable: Option<&P>,
-    ) -> Result<(GString, usize), glib::Error>;
-
-    #[cfg_attr(feature = "v2_56", deprecated)]
-    fn read_until_async<
-        P: IsA<Cancellable>,
-        Q: FnOnce(Result<(GString, usize), glib::Error>) + Send + 'static,
-    >(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-        cancellable: Option<&P>,
-        callback: Q,
-    );
-
-    #[cfg_attr(feature = "v2_56", deprecated)]
-
-    fn read_until_async_future(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-    ) -> Pin<Box_<dyn std::future::Future<Output = Result<(GString, usize), glib::Error>> + 'static>>;
-
-    fn read_upto<P: IsA<Cancellable>>(
-        &self,
-        stop_chars: &str,
-        cancellable: Option<&P>,
-    ) -> Result<(GString, usize), glib::Error>;
-
-    fn read_upto_async<
-        P: IsA<Cancellable>,
-        Q: FnOnce(Result<(GString, usize), glib::Error>) + Send + 'static,
-    >(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-        cancellable: Option<&P>,
-        callback: Q,
-    );
-
-    fn read_upto_async_future(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-    ) -> Pin<Box_<dyn std::future::Future<Output = Result<(GString, usize), glib::Error>> + 'static>>;
 
     fn set_byte_order(&self, order: DataStreamByteOrder);
 
@@ -362,188 +310,6 @@ impl<O: IsA<DataInputStream>> DataInputStreamExt for O {
                 Err(from_glib_full(error))
             }
         }
-    }
-
-    fn read_until<P: IsA<Cancellable>>(
-        &self,
-        stop_chars: &str,
-        cancellable: Option<&P>,
-    ) -> Result<(GString, usize), glib::Error> {
-        unsafe {
-            let mut length = mem::MaybeUninit::uninit();
-            let mut error = ptr::null_mut();
-            let ret = gio_sys::g_data_input_stream_read_until(
-                self.as_ref().to_glib_none().0,
-                stop_chars.to_glib_none().0,
-                length.as_mut_ptr(),
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                &mut error,
-            );
-            let length = length.assume_init();
-            if error.is_null() {
-                Ok((from_glib_full(ret), length))
-            } else {
-                Err(from_glib_full(error))
-            }
-        }
-    }
-
-    fn read_until_async<
-        P: IsA<Cancellable>,
-        Q: FnOnce(Result<(GString, usize), glib::Error>) + Send + 'static,
-    >(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-        cancellable: Option<&P>,
-        callback: Q,
-    ) {
-        let user_data: Box_<Q> = Box_::new(callback);
-        unsafe extern "C" fn read_until_async_trampoline<
-            Q: FnOnce(Result<(GString, usize), glib::Error>) + Send + 'static,
-        >(
-            _source_object: *mut gobject_sys::GObject,
-            res: *mut gio_sys::GAsyncResult,
-            user_data: glib_sys::gpointer,
-        ) {
-            let mut error = ptr::null_mut();
-            let mut length = mem::MaybeUninit::uninit();
-            let ret = gio_sys::g_data_input_stream_read_until_finish(
-                _source_object as *mut _,
-                res,
-                length.as_mut_ptr(),
-                &mut error,
-            );
-            let length = length.assume_init();
-            let result = if error.is_null() {
-                Ok((from_glib_full(ret), length))
-            } else {
-                Err(from_glib_full(error))
-            };
-            let callback: Box_<Q> = Box_::from_raw(user_data as *mut _);
-            callback(result);
-        }
-        let callback = read_until_async_trampoline::<Q>;
-        unsafe {
-            gio_sys::g_data_input_stream_read_until_async(
-                self.as_ref().to_glib_none().0,
-                stop_chars.to_glib_none().0,
-                io_priority.to_glib(),
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                Some(callback),
-                Box_::into_raw(user_data) as *mut _,
-            );
-        }
-    }
-
-    fn read_until_async_future(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-    ) -> Pin<Box_<dyn std::future::Future<Output = Result<(GString, usize), glib::Error>> + 'static>>
-    {
-        let stop_chars = String::from(stop_chars);
-        Box_::pin(crate::GioFuture::new(self, move |obj, send| {
-            let cancellable = Cancellable::new();
-            obj.read_until_async(&stop_chars, io_priority, Some(&cancellable), move |res| {
-                send.resolve(res);
-            });
-
-            cancellable
-        }))
-    }
-
-    fn read_upto<P: IsA<Cancellable>>(
-        &self,
-        stop_chars: &str,
-        cancellable: Option<&P>,
-    ) -> Result<(GString, usize), glib::Error> {
-        let stop_chars_len = stop_chars.len() as isize;
-        unsafe {
-            let mut length = mem::MaybeUninit::uninit();
-            let mut error = ptr::null_mut();
-            let ret = gio_sys::g_data_input_stream_read_upto(
-                self.as_ref().to_glib_none().0,
-                stop_chars.to_glib_none().0,
-                stop_chars_len,
-                length.as_mut_ptr(),
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                &mut error,
-            );
-            let length = length.assume_init();
-            if error.is_null() {
-                Ok((from_glib_full(ret), length))
-            } else {
-                Err(from_glib_full(error))
-            }
-        }
-    }
-
-    fn read_upto_async<
-        P: IsA<Cancellable>,
-        Q: FnOnce(Result<(GString, usize), glib::Error>) + Send + 'static,
-    >(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-        cancellable: Option<&P>,
-        callback: Q,
-    ) {
-        let stop_chars_len = stop_chars.len() as isize;
-        let user_data: Box_<Q> = Box_::new(callback);
-        unsafe extern "C" fn read_upto_async_trampoline<
-            Q: FnOnce(Result<(GString, usize), glib::Error>) + Send + 'static,
-        >(
-            _source_object: *mut gobject_sys::GObject,
-            res: *mut gio_sys::GAsyncResult,
-            user_data: glib_sys::gpointer,
-        ) {
-            let mut error = ptr::null_mut();
-            let mut length = mem::MaybeUninit::uninit();
-            let ret = gio_sys::g_data_input_stream_read_upto_finish(
-                _source_object as *mut _,
-                res,
-                length.as_mut_ptr(),
-                &mut error,
-            );
-            let length = length.assume_init();
-            let result = if error.is_null() {
-                Ok((from_glib_full(ret), length))
-            } else {
-                Err(from_glib_full(error))
-            };
-            let callback: Box_<Q> = Box_::from_raw(user_data as *mut _);
-            callback(result);
-        }
-        let callback = read_upto_async_trampoline::<Q>;
-        unsafe {
-            gio_sys::g_data_input_stream_read_upto_async(
-                self.as_ref().to_glib_none().0,
-                stop_chars.to_glib_none().0,
-                stop_chars_len,
-                io_priority.to_glib(),
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                Some(callback),
-                Box_::into_raw(user_data) as *mut _,
-            );
-        }
-    }
-
-    fn read_upto_async_future(
-        &self,
-        stop_chars: &str,
-        io_priority: glib::Priority,
-    ) -> Pin<Box_<dyn std::future::Future<Output = Result<(GString, usize), glib::Error>> + 'static>>
-    {
-        let stop_chars = String::from(stop_chars);
-        Box_::pin(crate::GioFuture::new(self, move |obj, send| {
-            let cancellable = Cancellable::new();
-            obj.read_upto_async(&stop_chars, io_priority, Some(&cancellable), move |res| {
-                send.resolve(res);
-            });
-
-            cancellable
-        }))
     }
 
     fn set_byte_order(&self, order: DataStreamByteOrder) {
