+++ src/lib.rs
@@ -29,6 +29,7 @@ extern crate futures_util;
 mod app_info;
 mod application;
 mod converter;
+mod data_input_stream;
 #[cfg(any(all(not(windows), not(target_os = "macos")), feature = "dox"))]
 mod desktop_app_info;
 mod error;
